--- app/src/main/java/nextstep/payments/data/PaymentCardsRepository.kt
@@ -1,13 +1,0 @@
-package nextstep.payments.data
-
-import nextstep.payments.domain.Card
-
-object PaymentCardsRepository {
-
-    private val _cards = mutableListOf<Card>()
-    val cards: List<Card> get() = _cards.toList()
-
-    fun addCard(card: Card) {
-        _cards.add(card)
-    }
-}
