+++ app/src/main/java/nextstep/payments/domain/PaymentCardsRepository.kt
@@ -1,0 +1,7 @@
+package nextstep.payments.domain
+
+interface PaymentCardsRepository {
+    fun getCards(): List<Card>
+
+    fun addCard(card: Card)
+}
