+++ app/src/main/java/nextstep/payments/screens/card/list/CardListViewModel.kt
@@ -5,16 +5,17 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
-import nextstep.payments.data.PaymentCardsRepository
+import nextstep.payments.data.InMemoryPaymentCardsRepository
+import nextstep.payments.domain.PaymentCardsRepository
 
 class CardListViewModel(
-    private val paymentCardsRepository: PaymentCardsRepository = PaymentCardsRepository,
+    private val paymentCardsRepository: PaymentCardsRepository = InMemoryPaymentCardsRepository,
 ): ViewModel() {
     private val _uiState: MutableStateFlow<CardListUiState> = MutableStateFlow(CardListUiState.Empty)
     val uiState: StateFlow<CardListUiState> = _uiState.asStateFlow()
 
     fun fetchCards() {
-        val cards = paymentCardsRepository.cards
+        val cards = paymentCardsRepository.getCards()
 
         if (cards.isEmpty()) {
             _uiState.update { CardListUiState.Empty }
